Add chrome extension
Chrome extension code to inject JS into spotify page to handle triggering of elements on the page once sockets tells it to.  Currently this just adds a button to the page that proves out the ability to click on the progress bar.  Need to figure out how to click on the index of the progress bar to set index of the song