Client prototype
Includes chrome plugin that can send a socket message from popup, accepts socket messages from background page.  Triggers on 'open' which will look for an open spotify page and navigate to a song to play, if no tab is open with spotify, it will open it for you and start playing.